+++ superwall/src/main/java/com/superwall/sdk/analytics/internal/Tracking.kt
@@ -50,6 +50,7 @@ suspend fun Superwall.track(event: Trackable): Result<TrackingResult> {
                 )
 
             dependencyContainer.delegateAdapter.handleSuperwallEvent(eventInfo = info)
+            emitSuperwallEvent(info)
 
             Logger.debug(
                 logLevel = LogLevel.debug,