+++ superwall/src/main/java/com/superwall/sdk/dependencies/DependencyContainer.kt
@@ -53,6 +53,9 @@ import com.superwall.sdk.paywall.presentation.internal.PresentationRequest
 import com.superwall.sdk.paywall.presentation.internal.PresentationRequestType
 import com.superwall.sdk.paywall.presentation.internal.request.PaywallOverrides
 import com.superwall.sdk.paywall.presentation.internal.request.PresentationInfo
+import com.superwall.sdk.paywall.presentation.rule_logic.cel.SuperscriptEvaluator
+import com.superwall.sdk.paywall.presentation.rule_logic.expression_evaluator.CombinedExpressionEvaluator
+import com.superwall.sdk.paywall.presentation.rule_logic.expression_evaluator.ExpressionEvaluating
 import com.superwall.sdk.paywall.presentation.rule_logic.javascript.DefaultJavascriptEvalutor
 import com.superwall.sdk.paywall.presentation.rule_logic.javascript.JavascriptEvaluator
 import com.superwall.sdk.paywall.request.PaywallRequest
@@ -82,6 +85,7 @@ import com.superwall.sdk.utilities.dateFormat
 import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
+import kotlinx.serialization.json.ClassDiscriminatorMode
 import kotlinx.serialization.json.Json
 import java.lang.ref.WeakReference
 import java.util.Base64
@@ -142,11 +146,30 @@ class DependencyContainer(
     private val ioScope
         get() = ioScope()
     private val evaluator by lazy {
-        DefaultJavascriptEvalutor(
-            ioScope = ioScope,
-            uiScope = uiScope,
-            context = context,
+        CombinedExpressionEvaluator(
             storage = storage,
+            factory = this,
+            evaluator =
+                DefaultJavascriptEvalutor(
+                    ioScope = ioScope,
+                    uiScope = uiScope,
+                    context = context,
+                    storage = storage,
+                ),
+            superscriptEvaluator =
+                SuperscriptEvaluator(
+                    json =
+                        Json(json()) {
+                            classDiscriminatorMode = ClassDiscriminatorMode.ALL_JSON_OBJECTS
+                            classDiscriminator = "type"
+                        },
+                    storage = storage.coreDataManager,
+                    factory = this,
+                    ioScope = ioScope,
+                ),
+            track = {
+                Superwall.instance.track(it)
+            },
         )
     }
 
@@ -624,7 +647,7 @@ class DependencyContainer(
 
     override suspend fun makeTriggers(): Set<String> = configManager.triggersByEventName.keys
 
-    override suspend fun provideJavascriptEvaluator(context: Context) = evaluator
+    override suspend fun provideRuleEvaluator(context: Context): ExpressionEvaluating = evaluator
 
     override fun makeConfigAttributes(): InternalSuperwallEvent.ConfigAttributes =
         InternalSuperwallEvent.ConfigAttributes(